# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Каталог товаров",
        "description": "Полноценное REST API для работы с продуктами",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://laravel.test.php-cat.com",
            "description": "сайт в сети для тестов"
        },
        {
            "url": "https://lara2602.local",
            "description": "Локальная площадка"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получить Bearer Token",
                "description": "Аутентификация пользователя и получение Bearer токена",
                "operationId": "ea134134deae11eb974c719450ef8a02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешная аутентификация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserWithTokenSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неверные учетные данные",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация нового пользователя",
                "description": "Создание нового пользователя и получение Bearer токена",
                "operationId": "58fe5a217ef88ae75e9b530a5e5817d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "success": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "data": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    },
                                    "message": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "code": {
                                        "type": "integer",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Успешная регистрация",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|randomtoken123"
                                                },
                                                "user": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "John Doe"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "user@example.com"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Успешная регистрация"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 201
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выход и отзыв токена",
                "description": "Отзывает текущий Bearer токен пользователя",
                "operationId": "5ebea01ab5f21df71d38f3505718a1d7",
                "responses": {
                    "200": {
                        "description": "Успешный выход",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Успешный выход"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Получить информацию о текущем пользователе",
                "description": "Возвращает информацию о текущем аутентифицированном пользователе",
                "operationId": "5d53d4426dbc772353e5adeb30e5bcb0",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "user@example.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Информация о пользователе"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "description": "Ошибка валидации",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Ошибка валидации"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "code": {
                        "type": "integer",
                        "example": 422
                    }
                },
                "type": "object"
            },
...
            "UnauthorizedError": {
                "description": "Ошибка авторизации",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    },
                    "errors": {
                        "type": "object",
                        "example": "{}"
                    }
                },
                "type": "object"
            },
...
            "ProductResource": {
                "description": "Продукт в ответе API",
                "required": [
                    "id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ноутбук HP"
                    },
                    "description": {
                        "type": "string",
                        "example": "Мощный ноутбук для работы",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 899.99
                    },
                    "price_formatted": {
                        "description": "Отформатированная цена",
                        "type": "string",
                        "example": "899.99"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "format": "uri",
                                "example": "http://api.example.com/api/products/1"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "description": "Пользователь в ответе API",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Петров"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user@example.com"
                    },
                    "email_verified_at": {
                        "description": "Дата подтверждения email",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00",
                        "nullable": true
                    },
                    "is_admin": {
                        "description": "Является ли пользователь администратором",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15 10:30:00"
                    },
                    "links": {
                        "properties": {
                            "self": {
                                "type": "string",
                                "format": "uri",
                                "example": "http://api.example.com/api/user"
                            },
                            "products": {
                                "description": "Ссылка на продукты пользователя",
                                "type": "string",
                                "format": "uri",
                                "example": "http://api.example.com/api/user/products"
                            },
                            "tokens": {
                                "description": "Ссылка на токены пользователя",
                                "type": "string",
                                "format": "uri",
                                "example": "http://api.example.com/api/user/tokens"
                            }
                        },
                        "type": "object"
                    },
                    "meta": {
                        "properties": {
                            "products_count": {
                                "description": "Количество продуктов пользователя",
                                "type": "integer",
                                "example": 15
                            },
                            "tokens_count": {
                                "description": "Количество активных токенов",
                                "type": "integer",
                                "example": 2
                            },
                            "last_login": {
                                "description": "Последний вход",
                                "type": "string",
                                "format": "date-time",
                                "example": "2024-01-20 14:25:00",
                                "nullable": true
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "UserWithTokenResource": {
                "description": "Пользователь с токеном в ответе API",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    {
                        "properties": {
                            "token": {
                                "description": "Bearer токен для аутентификации",
                                "type": "string",
                                "example": "1|randomtoken1234567890"
                            },
                            "token_expires_at": {
                                "description": "Срок действия токена",
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-01-15 10:30:00",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
            "UserWithTokenSuccessResponse": {
                "description": "Успешный ответ с пользователем и токеном",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserWithTokenResource"
                    },
                    "message": {
                        "type": "string",
                        "example": "Пользователь с токеном получен успешно"
                    },
                    "code": {
                        "type": "integer",
                        "example": 200
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Sanctum Token",
                "description": "Введите только сам токен (без слова Bearer). Пример: 1|abcdefghijklmnopqrstuvwxyz1234567890abcdef"
            }
        }
    },
...
}